+++ android/src/main/java/io/sentry/react/RNSentryModule.java
@@ -336,7 +336,10 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
                         String key = it.nextKey();
                         String value = otherUserKeys.getString(key);
 
-                        otherUserKeysMap.put(key, value);
+                        // other is ConcurrentHashMap and can't have null values
+                        if (value != null) {
+                            otherUserKeysMap.put(key, value);
+                        }
                     }
 
                     userInstance.setOthers(otherUserKeysMap);
@@ -388,13 +391,13 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
             }
 
             if (breadcrumb.hasKey("data")) {
-                ReadableMap data = breadcrumb.getMap("data");
-                ReadableMapKeySetIterator it = data.keySetIterator();
-                while (it.hasNextKey()) {
-                    String key = it.nextKey();
-                    String value = data.getString(key);
-
-                    breadcrumbInstance.setData(key, value);
+                final ReadableMap data = breadcrumb.getMap("data");
+                for(final Map.Entry<String, Object> entry : data.toHashMap().entrySet()) {
+                    final Object value = entry.getValue();
+                    // data is ConcurrentHashMap and can't have null values
+                    if (value != null) {
+                        breadcrumbInstance.setData(entry.getKey(), entry.getValue());
+                    }
                 }
             }