+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Make `configureScope` callback safe [#2510](https://github.com/getsentry/sentry-react-native/pull/2510)
 - Allows collecting app start and slow/frozen frames if Native SDK is inited manually [#2517](https://github.com/getsentry/sentry-react-native/pull/2517)
+- Nested breadcrumb data on android was not treated correctly [#2519](https://github.com/getsentry/sentry-react-native/pull/2519)
 
 ### Dependencies